+++ src/euroeval/dataset_configs/__init__.py
@@ -28,6 +28,7 @@ from .lithuanian import *  # noqa: F403
 from .norwegian import *  # noqa: F403
 from .polish import *  # noqa: F403
 from .portuguese import *  # noqa: F403
+from .romanian import *  # noqa: F403
 from .serbian import *  # noqa: F403
 from .slovak import *  # noqa: F403
 from .slovene import *  # noqa: F403