+++ src/scripts/constants.py
@@ -46,6 +46,7 @@ CHOICES_MAPPING = {
     "lv": "Izvēles",
     "nl": "Antwoordopties",
     "no": "Svaralternativer",
+    "ro": "Opțiuni",
     "pl": "Opcje",
     "pt": "Opções",
     "sk": "Možnosti",