+++ src/scripts/create_scala.py
@@ -45,6 +45,7 @@ from load_ud_pos import (
     load_nodt_nn_pos,
     load_pldt_pos,
     load_ptdt_pos,
+    load_rodt_pos,
     load_skdt_pos,
     load_sldt_pos,
     load_srdt_pos,
@@ -86,6 +87,7 @@ def main() -> None:
         "sl": load_sldt_pos,
         "hr": load_hrdt_pos,
         "hu": load_hudt_pos,
+        "ro": load_rodt_pos,
         "ca": load_cadt_pos,
     }