+++ src/scripts/create_winogrande.py
@@ -47,6 +47,7 @@ LANGUAGES = [
     "lv",
     "nl",
     "no",
+    "ro",
     "pl",
     "pt",
     "sk",